+++ HISTORY.md
@@ -67,6 +67,7 @@
 
 ### 2023-07-03
 * blocks: half-toned list symbols for collapsed RUN/CALL/LAUNCH/TELL/ASKL primitives
+* gui: added multi-sprite Parson's Puzzle capability
 
 ### 2023-07-02
 * new "Aleassa" costumes series, thanks, Meghan and Brian!